chore: update project structure - Added `entrypoints` directory - Moved `background.js` to `entrypoints` - Moved `popup.html`, `popup.css` and `popup.js` to `entrypoints/popup` - Updated `manifest.json`
