+++ entrypoints/popup/index.html
@@ -4,6 +4,7 @@
         <meta charset="UTF-8" />
         <meta http-equiv="X-UA-Compatible" content="IE=edge" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <meta name="manifest.type" content="browser_action" />
         <title>Kurozora</title>
 
         <!-- Fonts -->
@@ -11,10 +12,10 @@
         <link href="https://fonts.bunny.net/css?family=inter:100,200,300,400,500,600,700,800,900&display=swap" rel="stylesheet" />
 
         <!-- Styling -->
-        <link type="module" rel="stylesheet" href="./src/css/popup.css" />
+        <link type="module" rel="stylesheet" href="./style.css" />
 
         <!-- Scripts -->
-        <script type="module" src="./src/js/popup.js"></script>
+        <script type="module" src="./main.js"></script>
     </head>
 
     <body id="app"></body>
