+++ entrypoints/popup/main.js
